URI(undefined) now throws type error
Note that this differs from URI() which still works as a shortcut for
URI(location).

Since there is no reason to explicitly call URI with undefined, it is
safe to assume that this happens because of user error. URI now
correctly throws a type error with a message explaining that undefined
is not a valid argument.