Added more supporting methods for URN paths.
Use case: I work on some software that needs to accept URNs that may be
improperly formatted and encode them properly. This patch introduces a new
method, `recodeURNPath`, that will break the URN's path into segments and
make sure that each part of it is encoded properly.

Risks: URI.js treats "URN" as a catch-all category for everything that doesn't
look like an HTTP URL. Many of these schemes may or may not follow the URN
syntactical rules.